Everything compiling and passing again
Next step is to move in the Telemetry code
