+++ pkg/tesla/api/const.go
@@ -1,9 +1,5 @@
 package api
 
-const (
-	FleetTelemetryDataVersion = "fleet_telemetry/v1.0.0"
-)
-
 type TelemetryData struct {
 	Payloads [][]byte `json:"payloads"`
 }
